requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
asgiref=3.8.1=pypi_0
attrs=23.2.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.1.2=pypi_0
beautifulsoup4=4.12.3=pypi_0
build=1.2.1=pypi_0
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.3=pypi_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
chroma-hnswlib=0.7.3=pypi_0
chromadb=0.4.24=pypi_0
click=8.1.7=pypi_0
coloredlogs=15.0.1=pypi_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
dataclasses-json=0.6.4=pypi_0
datasets=2.19.0=pypi_0
deprecated=1.2.14=pypi_0
dill=0.3.8=pypi_0
dirtyjson=1.0.8=pypi_0
distro=1.9.0=pypi_0
faiss-cpu=1.8.0=pypi_0
fastapi=0.110.2=pypi_0
filelock=3.13.4=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.51.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
google-auth=2.29.0=pypi_0
googleapis-common-protos=1.63.0=pypi_0
greenlet=3.0.3=pypi_0
grpcio=1.62.2=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.22.2=pypi_0
humanfriendly=10.0=pypi_0
idna=3.7=pypi_0
importlib-metadata=7.0.0=pypi_0
importlib-resources=6.4.0=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.4.0=pypi_0
kiwisolver=1.4.5=pypi_0
kubernetes=29.0.0=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
llama-index=0.10.32=pypi_0
llama-index-agent-openai=0.2.3=pypi_0
llama-index-cli=0.1.12=pypi_0
llama-index-core=0.10.32=pypi_0
llama-index-embeddings-huggingface=0.2.0=pypi_0
llama-index-embeddings-openai=0.1.9=pypi_0
llama-index-graph-stores-neo4j=0.1.4=pypi_0
llama-index-indices-managed-llama-cloud=0.1.6=pypi_0
llama-index-legacy=0.9.48=pypi_0
llama-index-llms-openai=0.1.16=pypi_0
llama-index-multi-modal-llms-openai=0.1.5=pypi_0
llama-index-postprocessor-flag-embedding-reranker=0.1.2=pypi_0
llama-index-program-openai=0.1.6=pypi_0
llama-index-question-gen-openai=0.1.3=pypi_0
llama-index-readers-file=0.1.19=pypi_0
llama-index-readers-llama-parse=0.1.4=pypi_0
llama-index-vector-stores-chroma=0.1.6=pypi_0
llama-parse=0.4.2=pypi_0
llamaindex-py-client=0.1.19=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
marshmallow=3.21.1=pypi_0
matplotlib=3.8.4=pypi_0
mdurl=0.1.2=pypi_0
minijinja=1.0.16=pypi_0
mmh3=4.1.0=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
neo4j=5.20.0=pypi_0
nest-asyncio=1.6.0=pypi_0
networkx=3.3=pypi_0
nltk=3.8.1=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
onnxruntime=1.17.3=pypi_0
openai=1.23.6=pypi_0
openssl=3.0.13=h7f8727e_0
opentelemetry-api=1.24.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.24.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.24.0=pypi_0
opentelemetry-instrumentation=0.45b0=pypi_0
opentelemetry-instrumentation-asgi=0.45b0=pypi_0
opentelemetry-instrumentation-fastapi=0.45b0=pypi_0
opentelemetry-proto=1.24.0=pypi_0
opentelemetry-sdk=1.24.0=pypi_0
opentelemetry-semantic-conventions=0.45b0=pypi_0
opentelemetry-util-http=0.45b0=pypi_0
orjson=3.10.1=pypi_0
overrides=7.7.0=pypi_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
pillow=10.3.0=pypi_0
pip=23.3.1=py311h06a4308_0
posthog=3.5.0=pypi_0
protobuf=4.25.3=pypi_0
pulsar-client=3.5.0=pypi_0
pyarrow=16.0.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pydantic=2.7.1=pypi_0
pydantic-core=2.18.2=pypi_0
pyparsing=3.1.2=pypi_0
pypdf=4.2.0=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.0.0=pypi_0
python=3.11.9=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
redis=5.0.4=pypi_0
regex=2024.4.16=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=2.0.0=pypi_0
rich=13.7.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.4.2=pypi_0
scipy=1.13.0=pypi_0
sentence-transformers=2.7.0=pypi_0
setuptools=68.2.2=py311h06a4308_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.29=pypi_0
sqlite=3.41.2=h5eee18b_0
starlette=0.37.2=pypi_0
striprtf=0.0.26=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
threadpoolctl=3.4.0=pypi_0
tiktoken=0.6.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.19.1=pypi_0
torch=2.3.0=pypi_0
tqdm=4.66.2=pypi_0
transformers=4.40.1=pypi_0
triton=2.3.0=pypi_0
typer=0.12.3=pypi_0
typing-extensions=4.11.0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
uvicorn=0.29.0=pypi_0
uvloop=0.19.0=pypi_0
watchfiles=0.21.0=pypi_0
websocket-client=1.8.0=pypi_0
websockets=12.0=pypi_0
wheel=0.41.2=py311h06a4308_0
wrapt=1.16.0=pypi_0
xxhash=3.4.1=pypi_0
xz=5.4.6=h5eee18b_0
yarl=1.9.4=pypi_0
zipp=3.18.1=pypi_0
zlib=1.2.13=h5eee18b_0
